+++ hw/usb/dev-storage.c
@@ -348,7 +348,7 @@ static void usb_msd_handle_reset(USBDevice *dev)
 }
 
 static void usb_msd_handle_control(USBDevice *dev, USBPacket *p,
-               int request, int value, int index, int length, uint8_t *data)
+               int request, int value, int index, size_t length, uint8_t *data)
 {
     MSDState *s = (MSDState *)dev;
     SCSIDevice *scsi_dev;
